% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_duration.R
\name{check_time}
\alias{check_time}
\title{Check kobo interview time}
\usage{
check_time(data, duration_threshold_lower = 15,
  duration_threshold_upper = 100)
}
\arguments{
\item{data}{Dataframe with "start" and "end" columns.}

\item{duration_threshold_lower}{Minimum number of minutes to complete the form}

\item{duration_threshold_upper}{Maximum number of minutes to complete the form}
}
\value{
A dataframe with one row per potential issue. It has columns for the corresponding row index in the original data; the suspicious value (survey that is longer or shorter than it should be); the variable name in the original dataset in which the suspicious value occured; A description of the issue type.
}
\description{
Check kobo interview time
}
