% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{find_duplicates}
\alias{find_duplicates}
\title{Check the dataset for duplicates in a variable.}
\usage{
find_duplicates(data, duplicate.column.name)
}
\arguments{
\item{data}{a dataframe}

\item{duplicate.column.name}{the column in the dataframe}
}
\value{
A dataframe of the issue log format, containing the index and value of duplicated values
}
\description{
Check the dataset for duplicates in a variable.
}
\examples{
find_duplicates(data, "_uuid")
}
