% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{find_duplicates_uuid}
\alias{find_duplicates_uuid}
\title{Check the dataset for duplicates in the uuid variable.}
\usage{
find_duplicates_uuid(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
A dataframe of the issue log format, containing the index and value of duplicated uuids
}
\description{
Check the dataset for duplicates in the uuid variable.
}
