% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{find_other_responses}
\alias{find_other_responses}
\title{Check the dataset for variables}
\usage{
find_other_responses(data)
}
\arguments{
\item{data}{a dataframe}

\item{duplicate.column.name}{the column in the dataframe}
}
\value{
A dataframe of the issue log format, containing the index and value of other values that may need recoding
}
\description{
Check the dataset for variables
}
