% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_toUUID.R, R/inspection_add_data_columns.R
\name{index_toUUID}
\alias{index_toUUID}
\title{Replace indices by UUID to ease retrieval of records in dataset in cleaninginspectoR results}
\usage{
index_toUUID(inspectorResult, data, uuid.column.name = "uuid")

index_toUUID(inspectorResult, data, uuid.column.name = "uuid")
}
\arguments{
\item{inspectorResult}{a dataframe with cleaninginspectoR results}

\item{data}{a dataframe}

\item{uuid.column.name}{The name of the column containing the uuids. Default: "uuid".}

\item{inspection}{a dataframe with cleaninginspectoR results}

\item{df}{a dataframe}
}
\value{
A dataframe with the cleaninginspectoR results and an additional UUID column.

A dataframe with the cleaninginspectoR results and an additional UUID column.
}
\description{
Replace indices by UUID to ease retrieval of records in dataset in cleaninginspectoR results

Replace indices by UUID to ease retrieval of records in dataset in cleaninginspectoR results
}
\details{
Matches indices from cleaninginspectoR results with the UUIDs present in the data.

Matches indices from cleaninginspectoR results with the UUIDs present in the data.
}
