% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_all.R
\name{inspect_all}
\alias{inspect_all}
\title{Apply general data cleaning checks}
\usage{
inspect_all(df, uuid.column.name = NULL)
}
\arguments{
\item{uuid.column.name}{optional: The name of the column containing the uuids. If none is provided, will seach variable names for "uuid".}

\item{data}{a dataframe}
}
\value{
A dataframe with one row per potential issue, containing the row index, value and variable name in the original dataset, as well as a description of the issue type
}
\description{
Uses all other cleaning checks available in this package at once
}
\details{
for details see `?cleaninginspectoR`
}
\examples{
inspect_all(my_df)
}
