% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_all_csv_in_dir.R
\name{inspect_all_csv_in_dir}
\alias{inspect_all_csv_in_dir}
\title{run inspect_all() on all csv files in a folder and subfolders}
\usage{
inspect_all_csv_in_dir(source_dir = "./", pattern = "csv$",
  target_folder = "./", recursive = TRUE)
}
\arguments{
\item{source_dir}{folder to search for csv files that inspect_all() should be run on}

\item{pattern}{a regex pattern on which files in the folders to select. The default is "csv$"; this should be kept at the end of the pattern}

\item{target_folder}{path to the folder where the issue tables should be saved}

\item{recursive}{logical: if TRUE (default), also searches all subfolders of the current working directory}
}
\value{
a list of data frames with the outputs from each csv file in the current working directory (and subdirectories)#
}
\description{
run inspect_all() on all csv files in a folder and subfolders
}
