% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspection_add_data_columns.R
\name{inspection_add_data_columns}
\alias{inspection_add_data_columns}
\title{Add  column(s) to inspection results}
\usage{
inspection_add_data_columns(inspection, df, variables)
}
\arguments{
\item{inspection}{a dataframe with cleaninginspectoR results}

\item{df}{a dataframe containing the original data}

\item{variables}{The name(s) of the column(s) containing the variable(s) to add.}
}
\value{
A dataframe with the cleaninginspectoR results and an additional UUID column.
}
\description{
Add  column(s) to inspection results
}
\details{
Add columns from df to cleaninginspectoR results.
}
