% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{list_other_responses}
\alias{list_other_responses}
\title{list all responses in all columns that might be "specify other" responses to a multiple choice question}
\usage{
list_other_responses(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
A dataframe . It has columns for the corresponding row uuid in the original data; the suspicious value; the variable name in the original dataset in which the suspicious value occured.
}
\description{
list all responses in all columns that might be "specify other" responses to a multiple choice question
}
\details{
Performs a non-case sensitive search for "other" in english and french along the column names of the dataframe and returns all values.
}
