% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{sensitive_columns}
\alias{sensitive_columns}
\title{Check the dataset for senstive variables}
\usage{
sensitive_columns(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
A dataframe of the issue log format, containing the variables that may contain senstitive data
}
\description{
Check the dataset for senstive variables
}
