% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_checks.R
\name{find_outliers}
\alias{find_outliers}
\title{Find outliers in the data, deciding whether log or normal outlier functions should be used.}
\usage{
find_outliers(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
A list of numeric outliers that are 3 standard deviations from the mean
}
\description{
Find outliers in the data, deciding whether log or normal outlier functions should be used.
}
